requirements.txt
asgiref==3.3.1
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
Deprecated==1.2.13
Django==2.2.15
django-auth-ldap==2.2.0
django-cors-headers==3.2.1
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mysqlclient==1.3.13
packaging==20.4
Pillow==8.0.1
pip-review==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.0
pyparsing==2.4.7
python-ldap==3.2.0
pytz==2021.3
redis==4.0.2
requests==2.24.0
six==1.15.0
sqlparse==0.4.2
urllib3==1.25.11
uWSGI==2.0.20
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
